test: separated e2e tests
